+++ api/c/include/LIEF/PE/EnumToString.h
@@ -27,7 +27,6 @@ extern "C" {
 
 LIEF_API const char* PE_TYPES_to_string(enum LIEF_PE_PE_TYPES e);
 LIEF_API const char* MACHINE_TYPES_to_string(enum LIEF_PE_MACHINE_TYPES e);
-LIEF_API const char* SUBSYSTEM_to_string(enum LIEF_PE_SUBSYSTEM e);
 
 #ifdef __cplusplus
 }