+++ api/c/include/LIEF/PE/enums.h
@@ -16,6 +16,7 @@ extern "C" {
 #undef _LIEF_EN_2
 #undef _LIEF_EI
 
+
 #ifdef __cplusplus
 }
 #endif