+++ api/python/src/PE/objects/pyOptionalHeader.cpp
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 #include "PE/pyPE.hpp"
+#include "enums_wrapper.hpp"
 
 #include "LIEF/PE/OptionalHeader.hpp"
 
@@ -21,18 +22,51 @@
 #include <sstream>
 #include <nanobind/stl/string.h>
 #include <nanobind/stl/vector.h>
-#include <nanobind/stl/set.h>
 #include <nanobind/operators.h>
+
+#define PY_ENUM(x) to_string(x), x
 
 namespace LIEF::PE::py {
 
 template<>
 void create<OptionalHeader>(nb::module_& m) {
-  nb::class_<OptionalHeader, LIEF::Object>(m, "OptionalHeader",
+  nb::class_<OptionalHeader, Object> opt_header(m, "OptionalHeader",
       R"delim(
-      Class which represents the PE OptionalHeader structure..
-      )delim"_doc)
-    .def(nb::init<>())
+      Class which represents the PE OptionalHeader structure.
+      )delim"_doc);
+
+  enum_<OptionalHeader::SUBSYSTEM>(opt_header, "SUBSYSTEM")
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::UNKNOWN))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::NATIVE))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::WINDOWS_GUI))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::WINDOWS_CUI))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::OS2_CUI))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::POSIX_CUI))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::NATIVE_WINDOWS))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::WINDOWS_CE_GUI))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::EFI_APPLICATION))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::EFI_BOOT_SERVICE_DRIVER))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::EFI_RUNTIME_DRIVER))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::EFI_ROM))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::XBOX))
+    .value(PY_ENUM(OptionalHeader::SUBSYSTEM::WINDOWS_BOOT_APPLICATION));
+
+  enum_<OptionalHeader::DLL_CHARACTERISTICS>(opt_header, "DLL_CHARACTERISTICS", nb::is_arithmetic())
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::HIGH_ENTROPY_VA))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::DYNAMIC_BASE))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::FORCE_INTEGRITY))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::NX_COMPAT))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::NO_ISOLATION))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::NO_SEH))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::NO_BIND))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::APPCONTAINER))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::WDM_DRIVER))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::GUARD_CF))
+    .value(PY_ENUM(OptionalHeader::DLL_CHARACTERISTICS::TERMINAL_SERVER_AWARE));
+
+  opt_header
+    .def_static("create", OptionalHeader::create, "type"_a)
+
     .def_prop_rw("magic",
         nb::overload_cast<>(&OptionalHeader::magic, nb::const_),
         nb::overload_cast<PE_TYPE>(&OptionalHeader::magic),
@@ -193,7 +227,6 @@ void create<OptionalHeader>(nb::module_& m) {
         time, and any **DLL** that is loaded into a **critical** Windows process.
         )delim"_doc)
 
-
     .def_prop_ro("computed_checksum",
         &OptionalHeader::computed_checksum,
         R"delim(
@@ -206,38 +239,38 @@ void create<OptionalHeader>(nb::module_& m) {
 
     .def_prop_rw("subsystem",
         nb::overload_cast<>(&OptionalHeader::subsystem, nb::const_),
-        nb::overload_cast<SUBSYSTEM>(&OptionalHeader::subsystem),
+        nb::overload_cast<OptionalHeader::SUBSYSTEM>(&OptionalHeader::subsystem),
         R"delim(
-        Target subsystem (:class:`~lief.PE.SUBSYSTEM`) like Driver, XBox, Windows GUI, ..
+        Target subsystem (:class:`~lief.PE.OptionalHeader.SUBSYSTEM`) like Driver, XBox, Windows GUI, ..
         )delim"_doc)
 
     .def_prop_rw("dll_characteristics",
         nb::overload_cast<>(&OptionalHeader::dll_characteristics, nb::const_),
         nb::overload_cast<uint32_t>(&OptionalHeader::dll_characteristics),
         R"delim(
-        Some characteristics (:class:`~lief.PE.DLL_CHARACTERISTICS`) of the underlying binary like the support of the PIE.
+        Some characteristics (:class:`~lief.PE.OptionalHeader.DLL_CHARACTERISTICS`) of the underlying binary like the support of the PIE.
 
         The prefix ``dll`` comes from the official PE specifications but these characteristics
         are also used for **executables**
         )delim"_doc)
 
     .def("add",
-        nb::overload_cast<DLL_CHARACTERISTICS>(&OptionalHeader::add),
-        "Add the given " RST_CLASS_REF(lief.PE.DLL_CHARACTERISTICS) ""_doc,
+        nb::overload_cast<OptionalHeader::DLL_CHARACTERISTICS>(&OptionalHeader::add),
+        "Add the given " RST_CLASS_REF(lief.PE.OptionalHeader.DLL_CHARACTERISTICS) ""_doc,
         "characteristic"_a)
 
     .def("remove",
-        nb::overload_cast<DLL_CHARACTERISTICS>(&OptionalHeader::remove),
-        "Remove the given " RST_CLASS_REF(lief.PE.DLL_CHARACTERISTICS) ""_doc,
+        nb::overload_cast<OptionalHeader::DLL_CHARACTERISTICS>(&OptionalHeader::remove),
+        "Remove the given " RST_CLASS_REF(lief.PE.OptionalHeader.DLL_CHARACTERISTICS) ""_doc,
         "characteristic"_a)
 
     .def_prop_ro("dll_characteristics_lists",
         &OptionalHeader::dll_characteristics_list,
-        ":attr:`~lief.PE.OptionalHeader.dll_characteristics` as a list of " RST_CLASS_REF(lief.PE.DLL_CHARACTERISTICS) ""_doc)
+        ":attr:`~lief.PE.OptionalHeader.dll_characteristics` as a list of " RST_CLASS_REF(lief.PE.OptionalHeader.DLL_CHARACTERISTICS) ""_doc)
 
     .def("has",
-        nb::overload_cast<DLL_CHARACTERISTICS>(&OptionalHeader::has, nb::const_),
-        "``True`` if the given " RST_CLASS_REF(lief.PE.DLL_CHARACTERISTICS) " is in the "
+        nb::overload_cast<OptionalHeader::DLL_CHARACTERISTICS>(&OptionalHeader::has, nb::const_),
+        "``True`` if the given " RST_CLASS_REF(lief.PE.OptionalHeader.DLL_CHARACTERISTICS) " is in the "
         ":attr:`~lief.PE.OptionalHeader.dll_characteristics`"_doc,
         "characteristics"_a)
 
@@ -281,12 +314,12 @@ void create<OptionalHeader>(nb::module_& m) {
         nb::overload_cast<uint32_t>(&OptionalHeader::numberof_rva_and_size),
         "The number of " RST_CLASS_REF(lief.PE.DataDirectory) " that follow this header"_doc)
 
-    .def(nb::self += DLL_CHARACTERISTICS(), nb::rv_policy::reference_internal)
-    .def(nb::self -= DLL_CHARACTERISTICS(), nb::rv_policy::reference_internal)
+    .def(nb::self += OptionalHeader::DLL_CHARACTERISTICS(), nb::rv_policy::reference_internal)
+    .def(nb::self -= OptionalHeader::DLL_CHARACTERISTICS(), nb::rv_policy::reference_internal)
 
     .def("__contains__",
-        nb::overload_cast<DLL_CHARACTERISTICS>(&OptionalHeader::has, nb::const_),
-        "Check if the given " RST_CLASS_REF(lief.PE.DLL_CHARACTERISTICS) " is present"_doc)
+        nb::overload_cast<OptionalHeader::DLL_CHARACTERISTICS>(&OptionalHeader::has, nb::const_),
+        "Check if the given " RST_CLASS_REF(lief.PE.OptionalHeader.DLL_CHARACTERISTICS) " is present"_doc)
 
     LIEF_COPYABLE(OptionalHeader)
     LIEF_DEFAULT_STR(OptionalHeader);