+++ src/PE/OptionalHeader.cpp
@@ -13,281 +13,131 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <numeric>
-#include <iomanip>
 #include <algorithm>
 
 #include "LIEF/utils.hpp"
 
 #include "logging.hpp"
+#include "frozen.hpp"
 
 #include "LIEF/PE/hash.hpp"
 #include "LIEF/PE/OptionalHeader.hpp"
 #include "LIEF/PE/EnumToString.hpp"
 #include "PE/Structures.hpp"
 
+#include <spdlog/fmt/fmt.h>
 
 namespace LIEF {
 namespace PE {
 
+static constexpr std::array DLL_CHARACTERISTICS_LIST = {
+  OptionalHeader::DLL_CHARACTERISTICS::HIGH_ENTROPY_VA,
+  OptionalHeader::DLL_CHARACTERISTICS::DYNAMIC_BASE,
+  OptionalHeader::DLL_CHARACTERISTICS::FORCE_INTEGRITY,
+  OptionalHeader::DLL_CHARACTERISTICS::NX_COMPAT,
+  OptionalHeader::DLL_CHARACTERISTICS::NO_ISOLATION,
+  OptionalHeader::DLL_CHARACTERISTICS::NO_SEH,
+  OptionalHeader::DLL_CHARACTERISTICS::NO_BIND,
+  OptionalHeader::DLL_CHARACTERISTICS::APPCONTAINER,
+  OptionalHeader::DLL_CHARACTERISTICS::WDM_DRIVER,
+  OptionalHeader::DLL_CHARACTERISTICS::GUARD_CF,
+  OptionalHeader::DLL_CHARACTERISTICS::TERMINAL_SERVER_AWARE,
+};
+
 OptionalHeader::~OptionalHeader() = default;
 OptionalHeader& OptionalHeader::operator=(const OptionalHeader&) = default;
 OptionalHeader::OptionalHeader(const OptionalHeader&) = default;
-
-OptionalHeader::OptionalHeader() :
-  magic_{},
-  majorLinkerVersion_(9), // Arbitrary value
-  minorLinkerVersion_(0),
-  sizeOfCode_(0),
-  sizeOfInitializedData_(0),
-  sizeOfUninitializedData_(0),
-  addressOfEntryPoint_(0),
-  baseOfCode_(0),
-  baseOfData_(0),
-  imageBase_(0x00400000),
-  sectionAlignment_(0x1000),
-  fileAlignment_(0x200),
-  majorOperatingSystemVersion_(5), // Windows 2000
-  minorOperatingSystemVersion_(0),
-  majorImageVersion_(0),
-  minorImageVersion_(0),
-  majorSubsystemVersion_(5),
-  minorSubsystemVersion_(0),
-  win32VersionValue_(0),
-  sizeOfImage_(0),
-  sizeOfHeaders_(0),
-  checkSum_(0),
-  subsystem_(SUBSYSTEM::IMAGE_SUBSYSTEM_WINDOWS_CUI),
-  DLLCharacteristics_{},
-  sizeOfStackReserve_(0x200000),
-  sizeOfStackCommit_(0x1000),
-  sizeOfHeapReserve_(0x100000),
-  sizeOfHeapCommit_(0x1000),
-  loaderFlags_(0),
-  numberOfRvaAndSize_(details::DEFAULT_NUMBER_DATA_DIRECTORIES)
-{
-  sizeOfHeaders_ = sizeof(details::pe_dos_header) + sizeof(details::pe_header);
-  sizeOfHeaders_ = LIEF::align(sizeOfHeaders_, fileAlignment_);
-}
-
+OptionalHeader::OptionalHeader() = default;
 
 OptionalHeader::OptionalHeader(const details::pe32_optional_header& header):
   magic_(static_cast<PE_TYPE>(header.Magic)),
-  majorLinkerVersion_(header.MajorLinkerVersion),
-  minorLinkerVersion_(header.MinorLinkerVersion),
-  sizeOfCode_(header.SizeOfCode),
-  sizeOfInitializedData_(header.SizeOfInitializedData),
-  sizeOfUninitializedData_(header.SizeOfUninitializedData),
-  addressOfEntryPoint_(header.AddressOfEntryPoint),
-  baseOfCode_(header.BaseOfCode),
-  baseOfData_(header.BaseOfData),
-  imageBase_(header.ImageBase),
-  sectionAlignment_(header.SectionAlignment),
-  fileAlignment_(header.FileAlignment),
-  majorOperatingSystemVersion_(header.MajorOperatingSystemVersion),
-  minorOperatingSystemVersion_(header.MinorOperatingSystemVersion),
-  majorImageVersion_(header.MajorImageVersion),
-  minorImageVersion_(header.MinorImageVersion),
-  majorSubsystemVersion_(header.MajorSubsystemVersion),
-  minorSubsystemVersion_(header.MinorSubsystemVersion),
-  win32VersionValue_(header.Win32VersionValue),
-  sizeOfImage_(header.SizeOfImage),
-  sizeOfHeaders_(header.SizeOfHeaders),
-  checkSum_(header.CheckSum),
+  major_linker_version_(header.MajorLinkerVersion),
+  minor_linker_version_(header.MinorLinkerVersion),
+  sizeof_code_(header.SizeOfCode),
+  sizeof_initialized_data_(header.SizeOfInitializedData),
+  sizeof_uninitialized_data_(header.SizeOfUninitializedData),
+  entrypoint_(header.AddressOfEntryPoint),
+  baseof_code_(header.BaseOfCode),
+  baseof_data_(header.BaseOfData),
+  imagebase_(header.ImageBase),
+  section_align_(header.SectionAlignment),
+  file_align_(header.FileAlignment),
+  major_os_version_(header.MajorOperatingSystemVersion),
+  minor_os_version_(header.MinorOperatingSystemVersion),
+  major_image_version_(header.MajorImageVersion),
+  minor_image_version_(header.MinorImageVersion),
+  major_subsys_version_(header.MajorSubsystemVersion),
+  minor_subsys_version_(header.MinorSubsystemVersion),
+  win32_version_value_(header.Win32VersionValue),
+  sizeof_image_(header.SizeOfImage),
+  sizeof_headers_(header.SizeOfHeaders),
+  checksum_(header.CheckSum),
   subsystem_(static_cast<SUBSYSTEM>(header.Subsystem)),
-  DLLCharacteristics_(header.DLLCharacteristics),
-  sizeOfStackReserve_(header.SizeOfStackReserve),
-  sizeOfStackCommit_(header.SizeOfStackCommit),
-  sizeOfHeapReserve_(header.SizeOfHeapReserve),
-  sizeOfHeapCommit_(header.SizeOfHeapCommit),
-  loaderFlags_(header.LoaderFlags),
-  numberOfRvaAndSize_(header.NumberOfRvaAndSize)
+  dll_characteristics_(header.DLLCharacteristics),
+  sizeof_stack_reserve_(header.SizeOfStackReserve),
+  sizeof_stack_commit_(header.SizeOfStackCommit),
+  sizeof_heap_reserve_(header.SizeOfHeapReserve),
+  sizeof_heap_commit_(header.SizeOfHeapCommit),
+  loader_flags_(header.LoaderFlags),
+  nb_rva_size_(header.NumberOfRvaAndSize)
 {}
 
 OptionalHeader::OptionalHeader(const details::pe64_optional_header& header):
   magic_(static_cast<PE_TYPE>(header.Magic)),
-  majorLinkerVersion_(header.MajorLinkerVersion),
-  minorLinkerVersion_(header.MinorLinkerVersion),
-  sizeOfCode_(header.SizeOfCode),
-  sizeOfInitializedData_(header.SizeOfInitializedData),
-  sizeOfUninitializedData_(header.SizeOfUninitializedData),
-  addressOfEntryPoint_(header.AddressOfEntryPoint),
-  baseOfCode_(header.BaseOfCode),
-  baseOfData_(0), // Not in PE64
-  imageBase_(header.ImageBase),
-  sectionAlignment_(header.SectionAlignment),
-  fileAlignment_(header.FileAlignment),
-  majorOperatingSystemVersion_(header.MajorOperatingSystemVersion),
-  minorOperatingSystemVersion_(header.MinorOperatingSystemVersion),
-  majorImageVersion_(header.MajorImageVersion),
-  minorImageVersion_(header.MinorImageVersion),
-  majorSubsystemVersion_(header.MajorSubsystemVersion),
-  minorSubsystemVersion_(header.MinorSubsystemVersion),
-  win32VersionValue_(header.Win32VersionValue),
-  sizeOfImage_(header.SizeOfImage),
-  sizeOfHeaders_(header.SizeOfHeaders),
-  checkSum_(header.CheckSum),
+  major_linker_version_(header.MajorLinkerVersion),
+  minor_linker_version_(header.MinorLinkerVersion),
+  sizeof_code_(header.SizeOfCode),
+  sizeof_initialized_data_(header.SizeOfInitializedData),
+  sizeof_uninitialized_data_(header.SizeOfUninitializedData),
+  entrypoint_(header.AddressOfEntryPoint),
+  baseof_code_(header.BaseOfCode),
+  baseof_data_(0),
+  imagebase_(header.ImageBase),
+  section_align_(header.SectionAlignment),
+  file_align_(header.FileAlignment),
+  major_os_version_(header.MajorOperatingSystemVersion),
+  minor_os_version_(header.MinorOperatingSystemVersion),
+  major_image_version_(header.MajorImageVersion),
+  minor_image_version_(header.MinorImageVersion),
+  major_subsys_version_(header.MajorSubsystemVersion),
+  minor_subsys_version_(header.MinorSubsystemVersion),
+  win32_version_value_(header.Win32VersionValue),
+  sizeof_image_(header.SizeOfImage),
+  sizeof_headers_(header.SizeOfHeaders),
+  checksum_(header.CheckSum),
   subsystem_(static_cast<SUBSYSTEM>(header.Subsystem)),
-  DLLCharacteristics_(header.DLLCharacteristics),
-  sizeOfStackReserve_(header.SizeOfStackReserve),
-  sizeOfStackCommit_(header.SizeOfStackCommit),
-  sizeOfHeapReserve_(header.SizeOfHeapReserve),
-  sizeOfHeapCommit_(header.SizeOfHeapCommit),
-  loaderFlags_(header.LoaderFlags),
-  numberOfRvaAndSize_(header.NumberOfRvaAndSize)
+  dll_characteristics_(header.DLLCharacteristics),
+  sizeof_stack_reserve_(header.SizeOfStackReserve),
+  sizeof_stack_commit_(header.SizeOfStackCommit),
+  sizeof_heap_reserve_(header.SizeOfHeapReserve),
+  sizeof_heap_commit_(header.SizeOfHeapCommit),
+  loader_flags_(header.LoaderFlags),
+  nb_rva_size_(header.NumberOfRvaAndSize)
 {}
 
-PE_TYPE OptionalHeader::magic() const {
-  return magic_;
-}
 
+OptionalHeader OptionalHeader::create(PE_TYPE type) {
+  OptionalHeader header;
+  header.magic(type);
+  header.section_alignment(0x1000);
+  header.file_alignment(0x200);
+  header.numberof_rva_and_size(16);
+  header.subsystem(SUBSYSTEM::WINDOWS_CUI);
+  header.major_linker_version(9);
+  header.major_operating_system_version(6);
+  header.major_subsystem_version(6);
+  header.sizeof_stack_reserve(0x100000);
+  header.sizeof_stack_commit(0x1000);
+  header.sizeof_heap_reserve(0x100000);
+  header.sizeof_heap_commit(0x1000);
+  header.sizeof_headers(0x400);
 
-uint8_t OptionalHeader::major_linker_version() const {
-  return majorLinkerVersion_;
-}
-
-
-uint8_t OptionalHeader::minor_linker_version() const {
-  return minorLinkerVersion_;
-}
-
-
-uint32_t OptionalHeader::sizeof_code() const {
-  return sizeOfCode_;
-}
-
-
-uint32_t OptionalHeader::sizeof_initialized_data() const {
-  return sizeOfInitializedData_;
-}
-
-
-uint32_t OptionalHeader::sizeof_uninitialized_data() const {
-  return sizeOfUninitializedData_;
-}
-
-
-uint32_t OptionalHeader::addressof_entrypoint() const {
-  return addressOfEntryPoint_;
-}
-
-
-uint32_t OptionalHeader::baseof_code() const {
-  return baseOfCode_;
-}
-
-
-uint32_t OptionalHeader::baseof_data() const {
-  if (magic() == PE_TYPE::PE32_PLUS) {
-    LIEF_ERR("baseof_data is not present in PE64 binaries");
-    return 0;
+  if (type == PE_TYPE::PE32) {
+    header.imagebase(0x400000);
+  } else {
+    header.imagebase(0x140000000);
   }
-  return baseOfData_;
-}
-
-
-uint64_t OptionalHeader::imagebase() const {
-  return imageBase_;
-}
-
-
-uint32_t OptionalHeader::section_alignment() const {
-  return sectionAlignment_;
-}
-
-
-uint32_t OptionalHeader::file_alignment() const {
-  return fileAlignment_;
-}
-
-
-uint16_t OptionalHeader::major_operating_system_version() const {
-  return majorOperatingSystemVersion_;
-}
-
-
-uint16_t OptionalHeader::minor_operating_system_version() const {
-  return minorOperatingSystemVersion_;
-}
-
-
-uint16_t OptionalHeader::major_image_version() const {
-  return majorImageVersion_;
-}
-
-
-uint16_t OptionalHeader::minor_image_version() const {
-  return minorImageVersion_;
-}
-
-
-uint16_t OptionalHeader::major_subsystem_version() const {
-  return majorSubsystemVersion_;
-}
-
-
-uint16_t OptionalHeader::minor_subsystem_version() const {
-  return minorSubsystemVersion_;
-}
-
-
-uint32_t OptionalHeader::win32_version_value() const {
-  return win32VersionValue_;
-}
-
-
-uint32_t OptionalHeader::sizeof_image() const {
-  return sizeOfImage_;
-}
-
-
-uint32_t OptionalHeader::sizeof_headers() const {
-  return sizeOfHeaders_;
-}
-
-
-uint32_t OptionalHeader::checksum() const {
-  return checkSum_;
-}
-
-
-SUBSYSTEM OptionalHeader::subsystem() const {
-  return subsystem_;
-}
-
-
-uint32_t OptionalHeader::dll_characteristics() const {
-  return DLLCharacteristics_;
-}
-
-
-uint64_t OptionalHeader::sizeof_stack_reserve() const {
-  return sizeOfStackReserve_;
-}
-
-
-uint64_t OptionalHeader::sizeof_stack_commit() const {
-  return sizeOfStackCommit_;
-}
-
-
-uint64_t OptionalHeader::sizeof_heap_reserve() const {
-  return sizeOfHeapReserve_;
-}
-
-
-uint64_t OptionalHeader::sizeof_heap_commit() const {
-  return sizeOfHeapCommit_;
-}
-
-
-uint32_t OptionalHeader::loader_flags() const {
-  return loaderFlags_;
-}
-
-
-uint32_t OptionalHeader::numberof_rva_and_size() const {
-  return numberOfRvaAndSize_;
+  return header;
 }
 
 bool OptionalHeader::has(DLL_CHARACTERISTICS c) const {
@@ -302,176 +152,19 @@ void OptionalHeader::remove(DLL_CHARACTERISTICS c) {
   dll_characteristics(dll_characteristics() & (~ static_cast<uint32_t>(c)));
 }
 
+std::vector<OptionalHeader::DLL_CHARACTERISTICS> OptionalHeader::dll_characteristics_list() const {
+  std::vector<DLL_CHARACTERISTICS> characteristics;
+  characteristics.reserve(3);
+  std::copy_if(DLL_CHARACTERISTICS_LIST.begin(), DLL_CHARACTERISTICS_LIST.end(),
+               std::back_inserter(characteristics),
+               [this] (DLL_CHARACTERISTICS f) { return has(f); });
 
-std::set<DLL_CHARACTERISTICS> OptionalHeader::dll_characteristics_list() const {
-  std::set<DLL_CHARACTERISTICS> dll_charac;
-  std::copy_if(
-      std::begin(details::dll_characteristics_array), std::end(details::dll_characteristics_array),
-      std::inserter(dll_charac, std::begin(dll_charac)),
-      [this] (DLL_CHARACTERISTICS f) { return has(f); });
-
-  return dll_charac;
-}
-
-
-
-void OptionalHeader::magic(PE_TYPE magic) {
-  magic_ = static_cast<PE_TYPE>(magic);
-}
-
-
-void OptionalHeader::major_linker_version(uint8_t majorLinkerVersion) {
-  majorLinkerVersion_ = majorLinkerVersion;
-}
-
-
-void OptionalHeader::minor_linker_version(uint8_t minorLinkerVersion) {
-  minorLinkerVersion_ = minorLinkerVersion;
-}
-
-
-void OptionalHeader::sizeof_code(uint32_t sizeOfCode) {
-  sizeOfCode_ = sizeOfCode;
-}
-
-
-void OptionalHeader::sizeof_initialized_data(uint32_t sizeOfInitializedData) {
-  sizeOfInitializedData_ = sizeOfInitializedData;
-}
-
-
-void OptionalHeader::sizeof_uninitialized_data(uint32_t sizeOfUninitializedData) {
-  sizeOfUninitializedData_ = sizeOfUninitializedData;
-}
-
-
-void OptionalHeader::addressof_entrypoint(uint32_t addressOfEntryPoint) {
-  addressOfEntryPoint_ = addressOfEntryPoint;
-}
-
-
-void OptionalHeader::baseof_code(uint32_t baseOfCode) {
-  baseOfCode_ = baseOfCode;
-}
-
-
-void OptionalHeader::baseof_data(uint32_t baseOfData) {
-  if (magic() == PE_TYPE::PE32_PLUS) {
-    LIEF_ERR("baseof_data is not present in PE64 binaries");
-    return;
-  }
-  baseOfData_ = baseOfData;
-}
-
-
-void OptionalHeader::imagebase(uint64_t imageBase) {
-  imageBase_ = imageBase;
-}
-
-
-void OptionalHeader::section_alignment(uint32_t sectionAlignment) {
-  sectionAlignment_ = sectionAlignment;
-}
-
-
-void OptionalHeader::file_alignment(uint32_t fileAlignment) {
-  fileAlignment_ = fileAlignment;
-}
-
-
-void OptionalHeader::major_operating_system_version(uint16_t majorOperatingSystemVersion) {
-  majorOperatingSystemVersion_ = majorOperatingSystemVersion;
-}
-
-
-void OptionalHeader::minor_operating_system_version(uint16_t minorOperatingSystemVersion) {
-  minorOperatingSystemVersion_ = minorOperatingSystemVersion;
-}
-
-
-void OptionalHeader::major_image_version(uint16_t majorImageVersion) {
-  majorImageVersion_ = majorImageVersion;
-}
-
-
-void OptionalHeader::minor_image_version(uint16_t minorImageVersion) {
-  minorImageVersion_ = minorImageVersion;
-}
-
-
-void OptionalHeader::major_subsystem_version(uint16_t majorSubsystemVersion) {
-  majorSubsystemVersion_ = majorSubsystemVersion;
-}
-
-
-void OptionalHeader::minor_subsystem_version(uint16_t minorSubsystemVersion) {
-  minorSubsystemVersion_ = minorSubsystemVersion;
-}
-
-
-void OptionalHeader::win32_version_value(uint32_t win32VersionValue) {
-  win32VersionValue_ = win32VersionValue;
-}
-
-
-void OptionalHeader::sizeof_image(uint32_t sizeOfImage) {
-  sizeOfImage_ = sizeOfImage;
-}
-
-
-void OptionalHeader::sizeof_headers(uint32_t sizeOfHeaders) {
-  sizeOfHeaders_ = sizeOfHeaders;
-}
-
-
-void OptionalHeader::checksum(uint32_t checkSum) {
-  checkSum_ = checkSum;
-}
-
-
-void OptionalHeader::subsystem(SUBSYSTEM subsystem) {
-  subsystem_ = subsystem;
-}
-
-
-void OptionalHeader::dll_characteristics(uint32_t DLLCharacteristics) {
-  DLLCharacteristics_ = DLLCharacteristics;
-}
-
-
-void OptionalHeader::sizeof_stack_reserve(uint64_t sizeOfStackReserve) {
-  sizeOfStackReserve_ = sizeOfStackReserve;
-}
-
-
-void OptionalHeader::sizeof_stack_commit(uint64_t sizeOfStackCommit) {
-  sizeOfStackCommit_ = sizeOfStackCommit;
-}
-
-
-void OptionalHeader::sizeof_heap_reserve(uint64_t sizeOfHeapReserve) {
-  sizeOfHeapReserve_ = sizeOfHeapReserve;
-}
-
-
-void OptionalHeader::sizeof_heap_commit(uint64_t sizeOfHeapCommit) {
-  sizeOfHeapCommit_ = sizeOfHeapCommit;
-}
-
-
-void OptionalHeader::loader_flags(uint32_t loaderFlags) {
-  loaderFlags_ = loaderFlags;
-}
-
-
-void OptionalHeader::numberof_rva_and_size(uint32_t numberOfRvaAndSize) {
-  numberOfRvaAndSize_ = numberOfRvaAndSize;
+  return characteristics;
 }
 
 void OptionalHeader::accept(LIEF::Visitor& visitor) const {
   visitor.visit(*this);
 }
-
 
 OptionalHeader& OptionalHeader::operator+=(DLL_CHARACTERISTICS c) {
   add(c);
@@ -483,53 +176,83 @@ OptionalHeader& OptionalHeader::operator-=(DLL_CHARACTERISTICS c) {
   return *this;
 }
 
-
-
-
 std::ostream& operator<<(std::ostream& os, const OptionalHeader& entry) {
-  const auto& dll_chara = entry.dll_characteristics_list();
+  const std::vector<OptionalHeader::DLL_CHARACTERISTICS> list = entry.dll_characteristics_list();
+  std::vector<std::string> list_str;
+  list_str.reserve(list.size());
+  std::transform(list.begin(), list.end(), std::back_inserter(list_str),
+                 [] (const auto c) { return to_string(c); });
 
-  std::string dll_chara_str = std::accumulate(
-     std::begin(dll_chara),
-     std::end(dll_chara), std::string{},
-     [] (const std::string& a, DLL_CHARACTERISTICS b) {
-         return a.empty() ? std::string("\n- ") + to_string(b) : a + "\n- " + to_string(b);
-     });
-
-  os << std::hex;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Magic: "                          << static_cast<uint32_t>(entry.magic())                << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Major Linker Version: "           << static_cast<uint32_t>(entry.major_linker_version()) << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Minor Linker Version: "           << static_cast<uint32_t>(entry.minor_linker_version()) << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Code: "                   << entry.sizeOfCode_                                   << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Initialized Data: "       << entry.sizeOfInitializedData_                        << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Uninitialized Data: "     << entry.sizeOfUninitializedData_                      << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Address Of Entry Point: "         << entry.addressOfEntryPoint_                          << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Base Of Code: "                   << entry.baseOfCode_                                   << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Base Of Data: "                   << entry.baseOfData_                                   << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Image Base: "                     << entry.imageBase_                                    << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Section Alignment: "              << entry.sectionAlignment_                             << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "File Alignment: "                 << entry.fileAlignment_                                << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Major Operating System Version: " << entry.majorOperatingSystemVersion_                  << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Minor Operating System Version: " << entry.minorOperatingSystemVersion_                  << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Major Image Version: "            << entry.majorImageVersion_                            << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Minor Image Version: "            << entry.minorImageVersion_                            << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Major Subsystem Version: "        << entry.majorSubsystemVersion_                        << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Minor Subsystem Version: "        << entry.minorSubsystemVersion_                        << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Win32 Version Value: "            << entry.win32VersionValue_                            << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Image: "                  << entry.sizeOfImage_                                  << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Headers: "                << entry.sizeOfHeaders_                                << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "CheckSum: "                       << entry.checkSum_                                     << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Subsystem: "                      << to_string(entry.subsystem_)                         << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "DLL Characteristics: "            << dll_chara_str                                       << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Stack Reserve: "          << entry.sizeOfStackReserve_                           << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Stack Commit: "           << entry.sizeOfStackCommit_                            << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Heap Reserve: "           << entry.sizeOfHeapReserve_                            << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Size Of Heap Commit: "            << entry.sizeOfHeapCommit_                             << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Loader Flags: "                   << entry.loaderFlags_                                  << std::endl;
-  os << std::setw(33) << std::left << std::setfill(' ') << "Number Of RVA And Size: "         << entry.numberOfRvaAndSize_                           << std::endl;
-
+  os << fmt::format("Magic:                         0x{:x}\n", static_cast<uint32_t>(entry.magic()))
+     << fmt::format("Major/Minor linker version:    {}/{}\n", entry.major_linker_version(), entry.minor_linker_version())
+     << fmt::format("Size of code:                  0x{:x}\n", entry.sizeof_code())
+     << fmt::format("Size of initialized data:      0x{:x}\n", entry.sizeof_initialized_data())
+     << fmt::format("Size of uninitialized data:    0x{:x}\n", entry.sizeof_uninitialized_data())
+     << fmt::format("Entrypoint:                    0x{:x}\n", entry.addressof_entrypoint())
+     << fmt::format("Base of code:                  0x{:x}\n", entry.baseof_code())
+     << fmt::format("Base of data:                  0x{:x}\n", entry.baseof_data())
+     << fmt::format("Imagebase:                     0x{:x}\n", entry.imagebase())
+     << fmt::format("Section alignment:             0x{:x}\n", entry.section_alignment())
+     << fmt::format("File alignment:                0x{:x}\n", entry.file_alignment())
+     << fmt::format("Major/Minor operating system:  {}/{}\n", entry.major_operating_system_version(), entry.minor_operating_system_version())
+     << fmt::format("Major/Minor image version:     {}/{}\n", entry.major_image_version(), entry.minor_image_version())
+     << fmt::format("Major/Minor subsystem version: {}/{}\n", entry.major_subsystem_version(), entry.minor_subsystem_version())
+     << fmt::format("Win32 Version Value:           0x{:x}\n", entry.win32_version_value())
+     << fmt::format("Size of image:                 0x{:x}\n", entry.sizeof_image())
+     << fmt::format("Size of headers:               0x{:x}\n", entry.sizeof_headers())
+     << fmt::format("Checksum:                      0x{:x} (computed: 0x{:x})\n", entry.checksum(), entry.computed_checksum())
+     << fmt::format("Subsystem:                     {}\n", to_string(entry.subsystem()))
+     << fmt::format("Dll characteristics:           {}\n", fmt::join(list_str, ", "))
+     << fmt::format("Size of stack reserve:         0x{:x}\n", entry.sizeof_stack_reserve())
+     << fmt::format("Size of stack commit:          0x{:x}\n", entry.sizeof_stack_commit())
+     << fmt::format("Size of heap reserve:          0x{:x}\n", entry.sizeof_heap_reserve())
+     << fmt::format("Size of heap commit:           0x{:x}\n", entry.sizeof_heap_commit())
+     << fmt::format("Loader flags:                  0x{:x}\n", entry.loader_flags())
+     << fmt::format("Number of rva and size:        {}\n", entry.numberof_rva_and_size());
   return os;
 }
+
+const char* to_string(OptionalHeader::DLL_CHARACTERISTICS e) {
+  CONST_MAP(OptionalHeader::DLL_CHARACTERISTICS, const char*, 11) enumStrings {
+    { OptionalHeader::DLL_CHARACTERISTICS::HIGH_ENTROPY_VA,       "HIGH_ENTROPY_VA" },
+    { OptionalHeader::DLL_CHARACTERISTICS::DYNAMIC_BASE,          "DYNAMIC_BASE" },
+    { OptionalHeader::DLL_CHARACTERISTICS::FORCE_INTEGRITY,       "FORCE_INTEGRITY" },
+    { OptionalHeader::DLL_CHARACTERISTICS::NX_COMPAT,             "NX_COMPAT" },
+    { OptionalHeader::DLL_CHARACTERISTICS::NO_ISOLATION,          "NO_ISOLATION" },
+    { OptionalHeader::DLL_CHARACTERISTICS::NO_SEH,                "NO_SEH" },
+    { OptionalHeader::DLL_CHARACTERISTICS::NO_BIND,               "NO_BIND" },
+    { OptionalHeader::DLL_CHARACTERISTICS::APPCONTAINER,          "APPCONTAINER" },
+    { OptionalHeader::DLL_CHARACTERISTICS::WDM_DRIVER,            "WDM_DRIVER" },
+    { OptionalHeader::DLL_CHARACTERISTICS::GUARD_CF,              "GUARD_CF" },
+    { OptionalHeader::DLL_CHARACTERISTICS::TERMINAL_SERVER_AWARE, "TERMINAL_SERVER_AWARE" },
+  };
+  const auto it = enumStrings.find(e);
+  return it == enumStrings.end() ? "UNKNOWN" : it->second;
+}
+
+const char* to_string(OptionalHeader::SUBSYSTEM e) {
+  CONST_MAP(OptionalHeader::SUBSYSTEM, const char*, 14) enumStrings {
+    { OptionalHeader::SUBSYSTEM::UNKNOWN,                  "UNKNOWN" },
+    { OptionalHeader::SUBSYSTEM::NATIVE,                   "NATIVE" },
+    { OptionalHeader::SUBSYSTEM::WINDOWS_GUI,              "WINDOWS_GUI" },
+    { OptionalHeader::SUBSYSTEM::WINDOWS_CUI,              "WINDOWS_CUI" },
+    { OptionalHeader::SUBSYSTEM::OS2_CUI,                  "OS2_CUI" },
+    { OptionalHeader::SUBSYSTEM::POSIX_CUI,                "POSIX_CUI" },
+    { OptionalHeader::SUBSYSTEM::NATIVE_WINDOWS,           "NATIVE_WINDOWS" },
+    { OptionalHeader::SUBSYSTEM::WINDOWS_CE_GUI,           "WINDOWS_CE_GUI" },
+    { OptionalHeader::SUBSYSTEM::EFI_APPLICATION,          "EFI_APPLICATION" },
+    { OptionalHeader::SUBSYSTEM::EFI_BOOT_SERVICE_DRIVER,  "EFI_BOOT_SERVICE_DRIVER" },
+    { OptionalHeader::SUBSYSTEM::EFI_RUNTIME_DRIVER,       "EFI_RUNTIME_DRIVER" },
+    { OptionalHeader::SUBSYSTEM::EFI_ROM,                  "EFI_ROM" },
+    { OptionalHeader::SUBSYSTEM::XBOX,                     "XBOX" },
+    { OptionalHeader::SUBSYSTEM::WINDOWS_BOOT_APPLICATION, "WINDOWS_BOOT_APPLICATION" },
+  };
+  const auto it = enumStrings.find(e);
+  return it == enumStrings.end() ? "UNKNOWN" : it->second;
+}
+
+
+
 
 }
 }